Project ExploreWithMe: Main v0.1.2
<fix>: добавить в метод compilationToDto() класса CompilationMapper параметр Set<EventResponseDto> events;
<delete>: удалить из класса Event поле confirmedRequests;
<delete>: удалить из schema.sql основного модуля поле confirmed_requests в таблице events;
<fix>: заменить в DTO-классах и модели данных в основном модуле и schema.sql основного модуля типы полей confirmedRequests и participantLimit с Integer на Long;
<feat>: создать модель данных ConfirmedRequestsQuantity для получения количества подтверждённых заявок на событие из БД;
<fix>: убрать дублирование сохранения событий в классе RequestForEventService;
<feat>: добавить в репозиторий RequestForEventRepository методы countConfirmedRequestsByEvents() и countConfirmedRequestsBySingleEvent по получению объектов ConfirmedRequestsQuantity с количеством подтверждённых заявок;
<refactor>: исправить код в классах RequestForEventService, EventService, CompilationService, с учётом удаления из Event характеристики confirmedRequests и получению её из БД с помощью методов репозитория RequestForEventRepository;

+++ core/src/main/java/ru/practicum/explorewithme/dto/response/EventFullInfoResponseDto.java
@@ -16,7 +16,7 @@ import java.time.LocalDateTime;
 public class EventFullInfoResponseDto {
     private String annotation;
     private CategoryResponseDto category;
-    private Integer confirmedRequests;
+    private Long confirmedRequests;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss")
     private LocalDateTime createdOn;
     private String description;
@@ -26,7 +26,7 @@ public class EventFullInfoResponseDto {
     private InitiatorResponseDto initiator;
     private LocationResponseDto location;
     private Boolean paid;
-    private Integer participantLimit;
+    private Long participantLimit;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss")
     private LocalDateTime publishedOn;
     private Boolean requestModeration;
