+++ core/src/main/java/ru/practicum/explorewithme/dto/response/EventResponseDto.java
@@ -15,7 +15,7 @@ import java.time.LocalDateTime;
 public class EventResponseDto {
     private String annotation;
     private CategoryResponseDto category;
-    private Integer confirmedRequests;
+    private Long confirmedRequests;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss")
     private LocalDateTime eventDate;
     private Long id;
